UI screens header cleanup
