+++ UI/Background.h
@@ -10,3 +10,5 @@ void UIBackgroundInit(UIContext &dc);
 void UIBackgroundShutdown();
 void DrawGameBackground(UIContext &dc, const Path &gamePath, float x, float y, float z);
 void DrawBackground(UIContext &dc, float alpha, float x, float y, float z);
+
+uint32_t GetBackgroundColorWithAlpha(const UIContext &dc);
