+++ UI/CwCheatScreen.h
@@ -23,7 +23,7 @@
 #include "Common/UI/View.h"
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/Context.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 
 struct CheatFileInfo;
 class CWCheatEngine;
