+++ UI/DisplayLayoutScreen.h
@@ -21,9 +21,9 @@
 
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
+#include "Common/UI/PopupScreens.h"
 #include "GPU/Common/PostShader.h"
-
-#include "MiscScreens.h"
+#include "UI/MiscScreens.h"
 
 namespace UI {
 class ChoiceStrip;
