+++ UI/DriverManagerScreen.cpp
@@ -1,8 +1,11 @@
 #include "Common/File/VFS/ZipFileReader.h"
 #include "Common/Data/Format/JSONReader.h"
+#include "Common/Data/Text/I18n.h"
+#include "Common/System/Request.h"
 #include "Common/System/OSD.h"
 #include "Common/Log.h"
 #include "Common/StringUtils.h"
+#include "Common/UI/PopupScreens.h"
 
 #include "Core/Config.h"
 #include "Core/System.h"
@@ -11,6 +14,7 @@
 #include "UI/DriverManagerScreen.h"
 #include "UI/GameSettingsScreen.h"  // for triggerrestart
 #include "UI/OnScreenDisplay.h"
+#include "UI/MiscScreens.h"
 
 static Path GetDriverPath() {
 	if (g_Config.internalDataDirectory.empty()) {
