+++ UI/DriverManagerScreen.h
@@ -3,7 +3,7 @@
 #include "ppsspp_config.h"
 
 #include "Common/UI/UIScreen.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 #include "UI/TabbedDialogScreen.h"
 
 // Per-game settings screen - enables you to configure graphic options, control options, etc
