+++ UI/JitCompareScreen.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Common/UI/UIScreen.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 #include "UI/TabbedDialogScreen.h"
 
 class JitCompareScreen : public UITabbedBaseDialogScreen {
