+++ UI/MainScreen.h
@@ -24,7 +24,7 @@
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/ViewGroup.h"
 #include "Common/UI/TabHolder.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 #include "Common/File/PathBrowser.h"
 
 enum GameBrowserFlags {
