+++ UI/MiscScreens.h
@@ -124,30 +124,3 @@ protected:
 	void CreateViews() override;
 	const char *tag() const override { return "Credits"; }
 };
-
-class SettingInfoMessage : public UI::LinearLayout {
-public:
-	SettingInfoMessage(int align, float cutOffY, UI::AnchorLayoutParams *lp);
-
-	void Show(std::string_view text, const UI::View *refView = nullptr);
-
-	void Draw(UIContext &dc) override;
-	std::string GetText() const;
-
-private:
-	UI::TextView *text_ = nullptr;
-	double timeShown_ = 0.0;
-	float cutOffY_;
-	bool showing_ = false;
-};
-
-class ShinyIcon : public UI::ImageView {
-public:
-	ShinyIcon(ImageID atlasImage, UI::LayoutParams *layoutParams = 0) : UI::ImageView(atlasImage, "", UI::IS_DEFAULT, layoutParams) {}
-	void Draw(UIContext &dc) override;
-	void SetAnimated(bool anim) { animated_ = anim; }
-private:
-	bool animated_ = true;
-};
-
-uint32_t GetBackgroundColorWithAlpha(const UIContext &dc);
