+++ UI/RetroAchievementScreens.h
@@ -7,7 +7,7 @@
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/ViewGroup.h"
 #include "Core/RetroAchievements.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 #include "UI/TabbedDialogScreen.h"
 
 #include "ext/rcheevos/include/rc_client.h"
