+++ UI/TouchControlLayoutScreen.cpp
@@ -24,6 +24,7 @@
 #include "Common/Math/math_util.h"
 #include "Common/System/Display.h"
 #include "Common/UI/Context.h"
+#include "Common/UI/PopupScreens.h"
 
 #include "Common/CommonTypes.h"
 #include "Common/Log.h"
