+++ UI/TouchControlLayoutScreen.h
@@ -20,7 +20,7 @@
 #include "Common/UI/View.h"
 #include "Common/UI/ViewGroup.h"
 #include "Common/UI/TabHolder.h"
-#include "MiscScreens.h"
+#include "BaseScreens.h"
 
 class ControlLayoutView;
