+++ UI/UploadScreen.h
@@ -25,7 +25,7 @@
 
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/ViewGroup.h"
-#include "UI/MiscScreens.h"
+#include "UI/BaseScreens.h"
 #include "Common/TimeUtil.h"
 
 // Upload screen: Shows the user an ip address to go to in a web browser on the same network,
